Initial rewrite to expand item IDs from enums to integers
While this rewrite is functional and has no apparent major bugs, it is
incomplete. There are a number of TODOs added to the code to track some
outstanding work, but broadly:

  1. Mod hooks for items still need to be added.
  2. While TItem has been replaced, this is a strategy-layer struct; the
     tactical-layer structs (TWeapon, etc) have not been touched yet.
  3. There remain some disconnects where the new HL_TItem is not being
     used. Several fields, such as iBaseCharges and
     iCorpseToCharacterId, are populated but either used only in
     strategy or not used at all.

Since the current changes are functional and very large, I'm choosing to
commit and push as-is and return to finish the remaining work later.