Rewrite interceptions to be driven from the UI
The UI is now responsible for telling the interception state to update.
In return, the interception can now be aware of module usage and abort
attempts as soon as they occur. This avoids situations where the initial
simulation thought a ship would be dead/alive at a certain point, but
because the player used modules, that belief turns out to be mistaken.
In 1v1 ship battles this is insignificant, but in larger battles it can
cause ships to fire at dead targets, or ignore living ones.

Also adds temporary logic in LWCE_XGStrategyAI to remove a ship after it
is shot down, even though we can't currently spawn a mission for it.